requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dnspython==2.7.0
drf-spectacular==0.28.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.5
ollama==0.4.8
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pre_commit==4.2.0
pycodestyle==2.13.0
pydantic==2.11.4
pydantic-extra-types==2.10.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.4
rpds-py==0.24.0
ruff==0.11.7
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
threadpoolctl==3.6.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
websockets==15.0.1
